% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11naiveols-package.R
\name{naive_ols}
\alias{naive_ols}
\title{Naive OLS estimator}
\usage{
naive_ols(X, Y)
}
\arguments{
\item{X}{numeric matrix}

\item{Y}{numeric matrix}
}
\value{
numeric matrix
}
\description{
Naive OLS estimator
}
\examples{
X <- matrix(1, nrow = nrow(mtcars), ncol = 2)
X[, 2] <- mtcars$wt
Y <- matrix(mtcars$mpg, nrow = nrow(mtcars), ncol = 1)
naive_ols(X, Y)
}
